refactoring of auth server
